+++ libmtx/mtxfile/data.h
@@ -16,7 +16,7 @@
  * along with Libmtx.  If not, see <https://www.gnu.org/licenses/>.
  *
  * Authors: James D. Trotter <james@simula.no>
- * Last modified: 2022-01-14
+ * Last modified: 2022-04-14
  *
  * Matrix Market data lines.
  */
@@ -935,7 +935,7 @@ int mtxfiledata_set_constant_integer_double(
 int mtxfiledata_fread(
     union mtxfiledata * data,
     FILE * f,
-    int * lines_read,
+    int64_t * lines_read,
     int64_t * bytes_read,
     size_t line_max,
     char * linebuf,
@@ -973,7 +973,7 @@ int mtxfiledata_fread(
 int mtxfiledata_gzread(
     union mtxfiledata * data,
     gzFile f,
-    int * lines_read,
+    int64_t * lines_read,
     int64_t * bytes_read,
     size_t line_max,
     char * linebuf,
