+++ libmtx/mtxfile/mtxfile.h
@@ -16,7 +16,7 @@
  * along with Libmtx.  If not, see <https://www.gnu.org/licenses/>.
  *
  * Authors: James D. Trotter <james@simula.no>
- * Last modified: 2022-03-11
+ * Last modified: 2022-04-14
  *
  * Matrix Market files.
  */
@@ -586,7 +586,7 @@ int mtxfile_read(
     enum mtxprecision precision,
     const char * path,
     bool gzip,
-    int * lines_read,
+    int64_t * lines_read,
     int64_t * bytes_read);
 
 /**
@@ -610,7 +610,7 @@ int mtxfile_fread(
     struct mtxfile * mtxfile,
     enum mtxprecision precision,
     FILE * f,
-    int * lines_read,
+    int64_t * lines_read,
     int64_t * bytes_read,
     size_t line_max,
     char * linebuf);
@@ -638,7 +638,7 @@ int mtxfile_gzread(
     struct mtxfile * mtxfile,
     enum mtxprecision precision,
     gzFile f,
-    int * lines_read,
+    int64_t * lines_read,
     int64_t * bytes_read,
     size_t line_max,
     char * linebuf);
