+++ libmtx/mtxfile/size.c
@@ -16,7 +16,7 @@
  * along with Libmtx.  If not, see <https://www.gnu.org/licenses/>.
  *
  * Authors: James D. Trotter <james@simula.no>
- * Last modified: 2022-01-05
+ * Last modified: 2022-04-14
  *
  * Matrix Market size lines.
  */
@@ -356,7 +356,7 @@ static int freadline(
 int mtxfilesize_fread(
     struct mtxfilesize * size,
     FILE * f,
-    int * lines_read,
+    int64_t * lines_read,
     int64_t * bytes_read,
     size_t line_max,
     char * linebuf,
@@ -425,7 +425,7 @@ static int gzreadline(
 int mtxfilesize_gzread(
     struct mtxfilesize * size,
     gzFile f,
-    int * lines_read,
+    int64_t * lines_read,
     int64_t * bytes_read,
     size_t line_max,
     char * linebuf,
